requirements.txt
Flask==0.11.1
Flask-Cors==2.1.2
green==2.5.0
gunicorn==19.6.0
Pillow==3.2.0
requests==2.10.0
